requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-SocketIO==5.3.6
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
eventlet==0.36.1